Added OpenAL, things compile on windows.
Former-commit-id: ed01377eb17b49fb56c70342bdbde5aeb7780c6a
Former-commit-id: 7fdf3b4631ffc1bb24d52218a7b97081f2d586bf

+++ Sources/FlounderEngine/Helpers/HelperString.cpp
@@ -50,7 +50,7 @@ namespace Flounder
 
 	int HelperString::FindCharPos(const std::string &str, const char &c)
 	{
-		for (int i = 0; i < str.length(); i++)
+		for (int i = 0; i < static_cast<int>(str.length()); i++)
 		{
 			if (str.at(i) == c)
 			{
